refactor: extract document faker to crate
