+++ .github/package-filters/rs-packages.yml
@@ -65,3 +65,8 @@ dash-sdk:
   - packages/rs-sdk/**
   - *dapi_client
   - *drive
+
+random-document:
+  - .github/workflows/tests*
+  - packages/rs-random-document/**
+  - *dpp
